Separate place card component and styles for reusability
