+++ frontend/app/place/route.ts
@@ -1,0 +1,6 @@
+import { redirect } from 'next/navigation';
+import routes from '@/lib/constants/routes';
+
+export async function GET() {
+  redirect(routes.explore);
+}
