+++ frontend/components/generic/place-card/place-card.module.css
@@ -1,0 +1,44 @@
+.cardContent {
+  @apply relative flex flex-col gap-1.5 px-0 text-muted-foreground;
+
+  & .name {
+    @apply text-2xl font-semibold capitalize;
+  }
+
+  & .category,
+  & .location {
+    @apply text-base font-semibold capitalize;
+  }
+
+  & .location {
+    @apply -mt-1.5 mb-1.5;
+  }
+
+  & .description {
+    @apply text-xs text-primary;
+  }
+
+  & .image {
+    @apply h-48 w-full rounded-tl-[43px] rounded-tr-[43px] object-cover;
+  }
+
+  & .text {
+    @apply px-4 py-2 text-left;
+  }
+}
+
+.rating {
+  @apply absolute top-[42%] z-10 flex items-center gap-1;
+
+  & .background {
+    @apply flex h-7 w-12 rounded-br-xl rounded-tr-xl bg-background;
+  }
+
+  & .star {
+    @apply absolute left-1 z-20;
+  }
+
+  & .value {
+    @apply absolute left-5 mb-1 text-center text-muted-foreground;
+  }
+}
