+++ frontend/components/generic/place-card/place-card.tsx
@@ -1,9 +1,10 @@
 ﻿import Image from 'next/image';
+import { Place } from '@/lib/types/place';
 import { Card, CardContent } from '@/components/ui/card';
 import Star from '@/public/icons/star.svg';
-import { Place } from '@/lib/types/place';
-import styles from './interactive-showcase.module.css';
-import { CarouselItem } from '@/components/ui/carousel';
+import styles from './place-card.module.css';
+import routes from '@/lib/constants/routes';
+import Link from 'next/link';
 
 interface PlaceCardProps {
   place: Place;
@@ -11,14 +12,14 @@ interface PlaceCardProps {
 
 export default function PlaceCard({ place }: PlaceCardProps) {
   const pictures = place.pictures.filter(Boolean);
-  if (!pictures.length) return null;
+  if (!pictures.length) pictures.push(routes.placeholder);
 
   return (
-    <CarouselItem className={styles.carouselItem}>
+    <Link href={routes.place + '/' + place.id}>
       <Card>
         <CardContent className={styles.cardContent}>
-          <div className={styles.ratingContainer}>
-            <div className={styles.ratingBackground}></div>
+          <div className={styles.rating}>
+            <div className={styles.background}></div>
             <Image
               src={Star}
               alt={'Rating'}
@@ -26,7 +27,7 @@ export default function PlaceCard({ place }: PlaceCardProps) {
               height={11}
               className={styles.star}
             />
-            <p className={styles.rating}>{place.rating}</p>
+            <p className={styles.value}>{place.rating}</p>
           </div>
           <Image
             src={pictures[0]}
@@ -37,17 +38,15 @@ export default function PlaceCard({ place }: PlaceCardProps) {
             unoptimized
           />
           <div className={styles.text}>
-            <h3 className={styles.placeName}>{place.name}</h3>
-            <div className={styles.placeDescription}>
-              <p className={'capitalize'}>
-                {place.categoryId.replace(/_+/, ' ')}
-              </p>
-              <p>{place.location.address}</p>
-              <p>{place.description}</p>
-            </div>
+            <h3 className={styles.name}>{place.name}</h3>
+            <p className={styles.category}>
+              {place.categoryId.replace(/_+/, ' ')}
+            </p>
+            <p className={styles.location}>{place.location.address}</p>
+            <p className={styles.description}>{place.description}</p>
           </div>
         </CardContent>
       </Card>
-    </CarouselItem>
+    </Link>
   );
 }
